fix: merge #9796's Arc cache key with #9801, split regex.rs under the cap
