+++ crates/perry-runtime/src/regex/replace_expand.rs
@@ -5,6 +5,7 @@
 
 use super::replace_fn::{copy_replace_source, finish_replace_bytes};
 use super::*;
+use super::replace_expand_fancy::replace_regex_str_fancy;
 
 /// of `String.prototype.replace` special patterns that the Rust `regex`
 /// crate's own `$`-expansion does NOT cover: `$&` (matched substring),
